pyproject.toml
[project]
name = "bhu-text-mining"
version = "0.2.0"
description = ""
authors = [
    {name = "Kwang-Seuk", email = "kjeong@bhug.ac.kr"},
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "konlpy>=0.6.0",
    "requests>=2.32.3",
    "pillow>=11.1.0",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "networkx>=3.4.2",
    "llvmlite>=0.39.0",
    "numpy<2.1.0",
    "kss>=6.0.4",
    "mecab-python3>=1.0.10",
    "wordcloud>=1.9.4",
    "scikit-learn>=1.6.1",
    "hdbscan>=0.8.40",
    "bertopic>=0.16.4",
    "umap-learn>=0.5.7",
    "bert-score>=0.3.13",
    "reportlab>=4.4.4",
    "qrcode[pil]>=7.0",
    "pytesseract>=0.3.13",
    "opencv-python>=4.12.0.88",
    "pyyaml>=6.0",
    "anthropic>=0.25.0",
    "scipy>=1.11.0",
    "girth>=0.8.0",
    "stepmix>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pylint>=3.3.3",
    "black>=24.10.0",
]
jupyter = [
    "ipykernel>=6.29.5",
    "qtconsole>=5.6.1",
    "pyqt5>=5.15.11",
    "jupyter>=1.1.1",
]

[project.scripts]
bhu-exam = "src.cli:main"
bhu-ocr = "src.cli_ocr:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
